Relocate progress-line formatting helpers to actor_loop_flow
Move format_progress_line (61 LOC) and format_blocked_progress_line
(47 LOC) out of turn.rs into actor_loop_flow.rs. Their only
production consumers were the already-migrated actor loop methods.

Visibility upgrades on turn.rs items used by the migrated helpers:

- pub(super) on `tool_display` (Agent-free helper that builds
  ProgressDisplay)
- pub(super) on `ProgressDisplay` struct + 4 fields (action / path /
  note / status)

Imports added to actor_loop_flow.rs:

- super::progress_text::{paint, progress_available_width,
  format_progress_field, tool_color, tool_emoji, sanitize_for_progress}
- crate::modes::plan_act::PlanStage

The in-file `progress_tests` `#[cfg(test)] mod` in turn.rs gains
`use super::super::actor_loop_flow::{format_blocked_progress_line,
format_progress_line};` to keep the existing direct-call test
references valid; turn.rs prod imports of `tool_color` and `tool_emoji`
move to `#[cfg(test)] use super::progress_text::{...}` since their
only consumers are now tests.

Part of #680 / #681. Closes another round of back-import dependency
between actor_loop_flow.rs and turn.rs.
